requirements.txt
streamlit
pandas
matplotlib
vaderSentiment
google-api-python-client
python-dotenv
openpyxl
reportlab